switch TransitionMixCommand to code generation
Closes #37

+++ AtemSharp/State/Video/MixEffect/Transition/MixTransitionSettings.cs
@@ -8,5 +8,5 @@ public class MixTransitionSettings
     /// <summary>
     /// Rate of the mix transition in frames (0-250)
     /// </summary>
-    public int Rate { get; internal set; }
+    public byte Rate { get; internal set; }
 }
